work/python/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
captum>=0.6.0
shap>=0.44.0
lime>=0.2.0.1
river>=0.21.0
imagecorruptions>=1.1.2
scikit-learn>=1.3.0
scikit-image>=0.21.0
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.11.0
tqdm>=4.66.0
matplotlib>=3.7.0
wand
